Relocate health command
