--- mcc/health/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/poka-yoke/spaceflight/mcc/health/cmd"
-
-func main() {
-	cmd.Execute()
-}
